requirements.txt
opencv-python-headless==4.8.0.74
numpy
Pillow
torch
albumentations
efficientnet-pytorch
torchvision
